Update WAN Rest API with impact scope information [HZ-4594]
Some WAN API endpoints apply to the entire cluster, while some only apply to the receiving member. This commit adds notes to each endpoint specifying what the scope of its impact is.

I'm not sure if the `IMPORTANT:` boxes are the best way to present this, so please feel free to suggest better approaches!

Related to https://hazelcast.atlassian.net/browse/HZ-4594