Update capacity-planning.adoc
Added a few words to clarify the point that this calculation is only for the HybridLog requirement, and that the startup sanity check also includes the amount allocated at the data structure configuration level. It also clarifies the point that memory capacity planning includes hybridlog and the data itself - which should be obvious but isn't explicitly stated.