Remove hardcoded `Editions` content
The `Editions` page features a content listing of the various distributions,down to the exact filename of the JAR dependencies. This has not been maintained and become outdated.

The listing also doesn't explain what the contents _actually does_ - e.g. the case of the (empty) `custom-lib` directory.

There have been [discussions over the value of this content](https://hazelcast.slack.com/archives/C07066ELRRD/p1730463734768339), and as such updated it to provide a higher level summary.

Fixes: https://github.com/hazelcast/hz-docs/issues/1346