Add note about classpath for custom EPs in pipelines [HZG-287]
With the recently added support for UCN in Jet (available in next minor release), it is useful to mention this in documentation, particularly for the commonly used `EntryProcessor`. Even without UCN, the custom classes need to be on the classpath, so it's good to be explicit about this.

Fixes https://hazelcast.atlassian.net/browse/HZG-287