Client doc 294 (#1464)
See Docs ticket for details and SOW:
https://hazelcast.atlassian.net/browse/DOC-294

In summary, this docs work involves:

- moving some information from the Getting Started with Clients topic
into a new Clients & API Overview topic
- drafting a new Clients & API Overview topic that compares using
clients versus APIs
- tweaking the nav structure to create new Client and API sub-headings
(note the question I have about moving memcache under APIs rather than
clients)
- updating the Java client topic with more information about embedded
versus standalone Java client

This PR is now ready for review now that a decision has been made on how
to refer to Clients & APIs (see docs ticket).

Note that the client features graphic on the main HZC website may change
but, in the meantime, we will continue to link to it until a decision is
made about what to do with it.

---------

Co-authored-by: Oliver Howell <oliver.howell@hazelcast.com>