Some clarifications about data persistence [HZG-257] (#1510)
Some clarifications that data persistence is there to augment in-memory
partition backups and that planned cluster shutdowns must be done
cluster-wide. A rolling restart is not the same as a cluster shutdown,
it is a sequence of individual member restarts and so is more
appropriate to categorize as a speed improvement instead of a resiliency
one.