Remove more `full-version` usages (#1540)
`full-version` property is problematic as EE and OS are versioned
separately, which can lead to broken references.

_Where possible_ I've updated all `full-version` usages to be either OS
or EE, covering the following scenarios:
- Maven JAR version specifications (`pom.xml` or direct download links)
- `hazelcast-jet-kafka` link was just broken (previously missed because
the link isn't fully formed in the documentation)
- packaging references (brew etc)
- gradle references

*Many* of the above were producing the wrong result in our current
documentation (e.g. referencing `5.5.3` of Hazelcast OS in Maven which
doesn't exist).

One of the files also referenced a Jet artefact, but used the
`full-version` - Jet hasn't been updated since `4.x` so didn't work.
I've added a new property for the Jet version and have referenced that
instead, although the content is probably outdated and needs updating.