Fix Install Hazelcast OS JAR Download Link
In https://github.com/hazelcast/hz-docs/pull/1540 an error was introduced linking to the _EE_ version in the _OS_ docs in the case of _non-`SNAPSHOT`s. Not only was the link dead, but the EE version makes no sense in the OS installation steps.

This was trying to workaround the issue of lack of OS `SNAPSHOT`s, but actually this issue isn't resolvable - we can't link to what we don't publish (OS `SNAPSHOT`s) so instead the link should not be displayed in this case.