Update rolling upgrade MC version info
MC is not in lock step versions with Platform so it doesn't make sense to say "MC is compatible with the previous minor version". I tried to make it clear and accurate.

There should probably be a link on "Hazelcast Cluster Compatibility" (https://docs.hazelcast.com/management-center/5.7/getting-started/overview#hazelcast-cluster-compatibility), but either I'm not understanding the structure, or the "Preview" function doesn't show it.