Backport of Update map-connector.adoc (#1593)
Added note:

NOTE: Custom classes, such as an `EntryProcessor`, must be available to
all members of the cluster. This can be done by adding the classes to
each member's classpath directly, or by using {ucn}. If using {ucn},
ensure that the Job and underlying IMap are both associated with a
namespace that contains the same `EntryProcessor` implementation.