Backport of Update client-authorization.adoc (#1592)
Adds note:

All Hazelcast clients authorized to connect to a cluster are able to
receive proxy information for data structures present on the cluster.
This proxy only provides the client with the data structure's type and
name. Permissions for other actions related to these data structures can
be configured as required.