Clarify WAN queue capacity behaviour [HZG-308]
We only expose the primary events queue when we talk about "outbound queue size", which makes sense from a practical standpoint, but it can lead to confusion when you encounter scenarios such as partition promotions during migration. This can lead to metrics showing the WAN queue size exceeding the configured capacity, but it's only a shift of existing data, not a population of new data.

Without proper explanation, customers can be confused and concerned that the memory footprint of their WAN queues has suddenly increased, even though this is not the case.

This PR adds more information when discussing queue capacity in WAN, and also adds a bit more detail to the outbound queue metric description.

Fixes https://hazelcast.atlassian.net/browse/HZG-308