DOC-369 tutorials (#1522)
See Jira ticket: https://hazelcast.atlassian.net/browse/DOC-369

This work is to migrate the 6 HZC Clients buried at the foot of the
Tutorials page (see: https://docs.hazelcast.com/tutorials/), the content
for which resides in a different repo, and into the main Tech Docs to
make it easier for people to find.

A new client tutorials landing page has been created, beneath which are
the new tutorials which have been given a scrub and polish.

---------

Co-authored-by: Oliver Howell <oliver.howell@hazelcast.com>