[EMGR-481] Remove vert.x Hazelcast Enterprise Module Reference (#1721)
Hazelcast Enterprise Vert.x module was never published (and still
isn't). So, we need to update the documentation for now to remove its
reference as it's not possible for any user to follow.

Note: Updated tutorial should be tested -- I think it should be fine as
the Enterprise component was additive.