1782 faq increasing visibility (#1806)
For Docs ticket: https://github.com/hazelcast/hz-docs/issues/1782 (to
increase visibility of the individual FAQs)

Changes: used level-2-headings so that the individual questions appear
in the TOC on the right hand side, for ease of discovery/navigation.
Gave it a light copyedit/tidy. I've commented out the section line
breaks between questions for now as this was overkill IMHO but let me
know what you think.

Note that the section probably could benefit from some TLC later but for
this ticket, it's sufficient.