hz-docs issue 1785 - annotate Maven install config (#1815)
Follow on from https://github.com/hazelcast/hz-docs/pull/1811 to explain
the config.

Community Edition doesn't require specifying our repo because it's
hosted on Maven Central, but I don't think we need to explain that
difference between the two editions because we prompt users to choose
one or the other.

Not an entirely accurate use of in-line annotations but seemed the
neatest way to do it. Switched to includes to avoid duplication in
source.

https://github.com/hazelcast/hz-docs/issues/1785