1824-docs-feedback-for-advanced-features: improved description [v/5.5] (#1837)
Backport of https://github.com/hazelcast/hz-docs/pull/1832

For Docs ticket: https://github.com/hazelcast/hz-docs/issues/1824

The current documentation for the persist-wan-replicated-data setting
describes its effect in general terms, stating that when set to true,
incoming WAN events can be persisted to a database, but it does not
explicitly mention the relationship to MapStore or how this setting
interacts with MapStore processing.

The task is to improve the description to make this clearer. 

@jszap Please check you are happy with this new description. Thank you!

Co-authored-by: Amanda Lindsay <v-amanda.lindsay@hazelcast.com>