DOC-630: add security best practice [v/5.6] (#1976)
Backport of https://github.com/hazelcast/hz-docs/pull/1974

Add a security best practice topic, and improve the best practices
overview description to include this.

Please note that the other best practice topics will be scrubbed and the
style guide applied as part of a separate PR, so ignore the use of
initial caps etc. in their headings in the nav

Co-authored-by: Amanda Lindsay <v-amanda.lindsay@hazelcast.com>