Update helm-hazelcast-enterprise-chart.adoc (#2020)
The first section of doing a repo list, seeing it's github, then
rerunning the add read quite confusingly

Wording may not quite be right but how about simplifying it down?

Then at the end of the page - Notable changes - is it weird that the
latest version is last rather than first?

---------

Co-authored-by: Rob Swain <rob.swain@hazelcast.com>